+++ ft_atoi_bonus.c
@@ -6,17 +6,17 @@
 /*   By: houkaamo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/05 17:25:06 by houkaamo          #+#    #+#             */
-/*   Updated: 2026/02/05 17:25:17 by houkaamo         ###   ########.fr       */
+/*   Updated: 2026/02/07 16:21:57 by houkaamo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "checker_bonus.h"
 
-long long	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int			sign;
 	int			i;
-	long long	result;
+	long	result;
 
 	i = 0;
 	sign = 1;
@@ -34,8 +34,8 @@ long long	ft_atoi(const char *str)
 		if (sign == 1 && result > (INT_MAX - (str[i] - '0' )) / 10)
 			return ((long long)INT_MAX + 1);
 		if (sign == -1
-			&& result > (-(long long)INT_MIN - (str[i] - '0' )) / 10)
-			return ((long long)INT_MIN - 1);
+			&& result > (-(long)INT_MIN - (str[i] - '0' )) / 10)
+			return ((long)INT_MIN - 1);
 		result = result * 10 + (str[i++] - '0');
 	}
 	return (result * sign);
